pyproject.toml
[project]
name = "songpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.2",
    "pillow>=11.1.0",
    "streamlit>=1.41.1",
]
